Restructure config directory
