+++ src/main/java/de/unistuttgart/iste/meitrex/user_service/config/access_token/ExternalServiceProviderConfiguration.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.iste.meitrex.user_service.config;
+package de.unistuttgart.iste.meitrex.user_service.config.access_token;
 
 import de.unistuttgart.iste.meitrex.user_service.persistence.entity.ExternalServiceProvider;
 import lombok.Data;
